Moved get props function to interface
