+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 out/
 dist/
 code.js
+.DS_Store
