+++ manifest.json
@@ -1,5 +1,5 @@
 {
-  "name": "Figma-to-Rescript",
+  "name": "Figma to Code",
   "id": "1408123926034581384",
   "api": "1.0.0",
   "main": "dist/code.js",
